* process.c (proc_setgroups): fixed type.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@25034 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ process.c
@@ -4411,8 +4411,8 @@ proc_setgroups(VALUE obj, VALUE ary)
     Check_Type(ary, T_ARRAY);
 
     ngroups = RARRAY_LEN(ary);
-    if (ngroups > maxgroups)
-	rb_raise(rb_eArgError, "too many groups, %lu max", (unsigned long)maxgroups);
+    if (ngroups > (size_t)maxgroups)
+	rb_raise(rb_eArgError, "too many groups, %u max", maxgroups);
 
     groups = ALLOCA_N(rb_gid_t, ngroups);
